code-agent-refactored/pyproject.toml
[project]
name = "code-agent"
version = "0.2.0"
description = "A local AI coding agent powered by Groq with filesystem and execution tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=1.2.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
code-agent = "main:main"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[tool.ruff]
line-length = 100
target-version = "py311"
